Use NES resolution
After add a Mario Bros sprite at 800X600 resolution, this not fit very well,
this is so small. So I decided to follow the NES resolution.

Since the resolution is too small, I resized to 2x.

We have a plugin[1] and a tutorial[2] after some research and some tests. I
decided to do this via CSS, since I do not see any diference using the plugin
and using CSS. I tested only a Mario sprite on the screen, and do not search
pixel by pixel only a visual simple test.

I think that following this idea I am doing something like the emulators do,
when you resize the screen.

[1]: https://github.com/beeglebug/phaser-pixel-scaler
[2]: http://www.photonstorm.com/phaser/pixel-perfect-scaling-a-phaser-game