Add a small pause before load the game
When starts the game the BMG waits around 3 seconds before starting.

Adding a small pause on the preload the sounds start almost the same time of the
game. This is not good for a production environment, since the main idea is to
run this local, I do not need to care about this for now.

With that It is possible to see the logo, in development environment the loading
is so fast that is so hard to see the logo.